refactor: separated out some impl details
