+++ src/utils/helpers.ts
@@ -1,4 +1,4 @@
-import type { LocationPoint, SpeedDataType } from "../types";
+import type { TravelDataType, UserDataType } from "../types";
 
 export const shortenStr = (str: string, charCount: number) => {
   if (str.length <= charCount) return;
@@ -27,38 +27,6 @@ export const fetchData = async (URL: string, hasData: boolean, reqData?: any) =>
   }
 };
 
-export const processData = (locHist: LocationPoint[]) => {
-  return locHist.map((pt) => ({ time: pt.ts, speed: pt.speed || 0 }));
-};
-
-export const graphData = (data: SpeedDataType[]) => {
-  const secsData = data.map((o, i, data) => ({ ...o, time: (o.time - data[0].time) / 1000 }));
-
-  const leapCount = Math.floor(secsData.length / 20);
-  return leapCount < 1
-    ? secsData
-    : secsData.reduce((acc, v, i) => {
-        if (i % leapCount !== 0) return acc;
-        return [...acc, v];
-      }, [] as SpeedDataType[]);
-};
-
-export const getTestGraphData = () => {
-  let data: SpeedDataType[] = [];
-  for (let i = 0; i < 10; i++) data.push({ time: i * 0.5, speed: 15 + Math.floor(Math.random() * 10) });
-
-  return data;
-};
-
-export const getDist = (sData: SpeedDataType[]) => {
-  let dist = 0;
-  for (let i = 1; i < sData.length; i++) {
-    dist += ((sData[i].speed + sData[i - 1].speed) / 2) * (sData[i].time - sData[i - 1].time);
-  }
-
-  return dist;
-};
-
 export const getBals = async (userAcc: string) => {
   return fetchData("api/getBalances", true, { accId: userAcc })
     .then((balInfo) => {
@@ -68,4 +36,38 @@ export const getBals = async (userAcc: string) => {
       console.log(err);
       return { hbar: "", token: 0 };
     });
+};
+
+export const getDataToSave = (userData: UserDataType, travelData: TravelDataType) => ({
+  accID: userData.userAcc,
+  vehData: {
+    modelConfig: travelData.currVehicle?.modelConfig,
+    engineConfig: travelData.currVehicle?.engineConfig,
+    fuel: travelData.currVehicle?.fuel,
+    co2pm: travelData.currVehicle?.co2pm,
+  },
+  speedData: travelData.speedData,
+});
+
+export const pushDataReq = (dataToSave: any) => {
+  return fetchData("api/pushData", true, dataToSave)
+    .then((rdata) => {
+      if (rdata.error) {
+        console.log(rdata.message);
+        return "Error saving data";
+      } else {
+        return "Data Saved!";
+      }
+    })
+    .catch((err) => {
+      console.log(err);
+      return "Error saving data";
+    });
+};
+
+export const calcEmissions = (travelData: TravelDataType) => {
+  if (travelData.currVehicle && travelData.currVehicle.co2pm)
+    return (travelData.currVehicle.co2pm * travelData.distance) / 1609;
+
+  return 0;
 };
